refactor: extract searchKey checkbox filtering
